update : 프론트 WebConfig 생성
